manager_bot/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
requests>=2.31